+++ src/components/landing/Content.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import Button from "@/components/common/Button";
+import Button from "@/components/common/Button/Button";
 
 export default function Content() {
   return (
