+++ src/lib/editMockData.tsx
@@ -1,0 +1,43 @@
+export const Mock_MyData = {
+    id: 657,
+    email: "dex@naver.com",
+    nickname: "김진영",
+    profileImageUrl: "",
+    createdAt: "2024-01-29T17:38:10.326Z",
+    updatedAt: "2024-01-29T17:38:10.326Z",
+    };
+    
+export const Mock_DashBoard = [
+    {
+    id: 2564,
+    title: "중급 프로젝트",
+    color: "#7ac555",
+    userId: 588,
+    createdAt: "2024-01-26T22:22:53.928Z",
+    updatedAt: "2024-01-26T22:22:53.928Z",
+    createdByMe: true,
+    }
+];
+
+export const Mock_Members = [
+    {
+      id: 2,
+      userId: 2,
+      email: "member1@example.com",
+      nickname: "토게피",
+      profileImageUrl: "",
+      createdAt: "2024-01-30T13:07:45.642Z",
+      updatedAt: "2024-01-30T13:07:45.642Z",
+      isOwner: false,
+    },
+    {
+      id: 3,
+      userId: 3,
+      email: "member2@example.com",
+      nickname: "망나뇽",
+      profileImageUrl: "",
+      createdAt: "2024-01-30T13:07:45.642Z",
+      updatedAt: "2024-01-30T13:07:45.642Z",
+      isOwner: false,
+    },
+  ];
